external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/glsl/misc/shader-with-undefined-preprocessor-symbol.frag-expected.txt

CONSOLE MESSAGE: line 24: *** Error compiling shader '[object WebGLShader]':ERROR: 0:2: '' : incorrect preprocessor directive 
WARNING: 0:2: '' : unexpected tokens following #if preprocessor directive - expected a newline 
ERROR: 0:10: 'unexpected EOF' : syntax error 

fragment shader (3.4) undefined preprocessor symbol should fail

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

test: fragment shader (3.4) undefined preprocessor symbol should fail
default vertex shader
test fragment shaderPASS fragment shader (3.4) undefined preprocessor symbol should fail
PASS successfullyParsed is true

TEST COMPLETE
